Terraform: Dedicated network for development
Move the development machine off of the default network and onto a network
dedicated to development purposes.  This allows for firewall rules (like
allowing development HTTP traffic) that are isolated from the default
network.  This is all surprisingly easier than it is on AWS, as GCP provides
useful default routes and routers.

This also updates the SSH public key to the one I'm currently using because
I'm sick of stripping it out.  Finally, I am giving in and recognizing that
I will never step down from a e2-standard-2 instance to a shared-core
instance in the foreseeable future.